requirements.txt
langchain
langchain-openai
langchain-google-genai
python-dotenv
requests
faiss-cpu
numpy
markitdown[pdf]
tqdm
Pillowexit
